datarush_hackathon/agents/extensions/data_analysis_agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-analysis-agent"
version = "0.1.0"
description = "Data Analysis Agent for DataRush Hackathon"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-adk",
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "plotly>=5.0.0",
    "streamlit>=1.28.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=5.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["data_analysis_agent*"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

